refactor: split multi_provider.py into a package with _strategies and _provider modules
