+++ openfeature/provider/multi_provider/_provider.py
@@ -34,22 +34,16 @@ from openfeature.provider import (
 )
 from openfeature.track import TrackingEventDetails
 
-__all__ = [
-    "ComparisonStrategy",
-    "EvaluationStrategy",
-    "FirstMatchStrategy",
-    "FirstSuccessfulStrategy",
-    "MultiProvider",
-    "ProviderEntry",
-]
+from ._strategies import (
+    EvaluationStrategy,
+    FirstMatchStrategy,
+    RunMode,
+    T,
+    _ProviderEvaluation,
+    _build_aggregated_error,
+)
 
 logger = logging.getLogger("openfeature")
-
-T = typing.TypeVar("T", bound=FlagValueType)
-RunMode: typing.TypeAlias = typing.Literal["sequential", "parallel"]
-ComparisonMismatchHandler: typing.TypeAlias = Callable[
-    [str, Mapping[str, FlagResolutionDetails[FlagValueType]]], None
-]
 
 
 @dataclass(frozen=True)
@@ -59,293 +53,10 @@ class ProviderEntry:
 
 
 @dataclass(frozen=True)
-class _ProviderEvaluation(typing.Generic[T]):
-    provider_name: str
-    provider: FeatureProvider
-    result: FlagResolutionDetails[T]
-
-
-@dataclass(frozen=True)
 class _ProviderHookRuntime:
     flag_type: FlagType
     client_metadata: typing.Any
     hook_hints: HookHints
-
-
-class EvaluationStrategy(typing.Protocol):
-    run_mode: RunMode
-
-    def should_use_result(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool: ...
-
-    def should_continue(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool: ...
-
-    def determine_final_result(
-        self,
-        flag_key: str,
-        default_value: FlagValueType,
-        evaluations: list[_ProviderEvaluation[FlagValueType]],
-    ) -> FlagResolutionDetails[FlagValueType]: ...
-
-
-def _is_success(result: FlagResolutionDetails[FlagValueType]) -> bool:
-    return result.error_code is None and result.reason != Reason.ERROR
-
-
-def _validate_run_mode(run_mode: RunMode) -> RunMode:
-    if run_mode not in ("sequential", "parallel"):
-        raise ValueError(f"Unsupported run_mode '{run_mode}'")
-    return run_mode
-
-
-def _format_result_error(
-    provider_name: str, result: FlagResolutionDetails[FlagValueType]
-) -> str:
-    error_code = (
-        result.error_code.value if result.error_code else ErrorCode.GENERAL.value
-    )
-    error_message = result.error_message or "Unknown error"
-    return f"{provider_name}: {error_code} ({error_message})"
-
-
-def _build_aggregated_error(
-    flag_key: str,
-    default_value: FlagValueType,
-    evaluations: list[_ProviderEvaluation[FlagValueType]],
-    prefix: str,
-) -> FlagResolutionDetails[FlagValueType]:
-    if not evaluations:
-        return FlagResolutionDetails(
-            value=default_value,
-            reason=Reason.ERROR,
-            error_code=ErrorCode.GENERAL,
-            error_message=f"{prefix} for flag '{flag_key}': no providers returned a result",
-        )
-
-    errors_text = "; ".join(
-        _format_result_error(evaluation.provider_name, evaluation.result)
-        for evaluation in evaluations
-    )
-    return FlagResolutionDetails(
-        value=default_value,
-        reason=Reason.ERROR,
-        error_code=ErrorCode.GENERAL,
-        error_message=f"{prefix} for flag '{flag_key}': {errors_text}",
-    )
-
-
-class FirstMatchStrategy:
-    def __init__(self, run_mode: RunMode = "sequential") -> None:
-        self.run_mode = _validate_run_mode(run_mode)
-
-    def should_use_result(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool:
-        del flag_key
-        del provider_name
-        return _is_success(result)
-
-    def should_continue(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool:
-        del flag_key
-        del provider_name
-        return result.error_code == ErrorCode.FLAG_NOT_FOUND
-
-    def determine_final_result(
-        self,
-        flag_key: str,
-        default_value: FlagValueType,
-        evaluations: list[_ProviderEvaluation[FlagValueType]],
-    ) -> FlagResolutionDetails[FlagValueType]:
-        for evaluation in evaluations:
-            if self.should_use_result(
-                flag_key, evaluation.provider_name, evaluation.result
-            ):
-                return evaluation.result
-            if not self.should_continue(
-                flag_key, evaluation.provider_name, evaluation.result
-            ):
-                return evaluation.result
-        if evaluations:
-            return evaluations[-1].result
-        return _build_aggregated_error(
-            flag_key,
-            default_value,
-            evaluations,
-            "Multi-provider evaluation failed",
-        )
-
-
-class FirstSuccessfulStrategy:
-    def __init__(self, run_mode: RunMode = "sequential") -> None:
-        self.run_mode = _validate_run_mode(run_mode)
-
-    def should_use_result(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool:
-        del flag_key
-        del provider_name
-        return _is_success(result)
-
-    def should_continue(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool:
-        del flag_key
-        del provider_name
-        del result
-        return True
-
-    def determine_final_result(
-        self,
-        flag_key: str,
-        default_value: FlagValueType,
-        evaluations: list[_ProviderEvaluation[FlagValueType]],
-    ) -> FlagResolutionDetails[FlagValueType]:
-        for evaluation in evaluations:
-            if _is_success(evaluation.result):
-                return evaluation.result
-        return _build_aggregated_error(
-            flag_key,
-            default_value,
-            evaluations,
-            "All providers failed",
-        )
-
-
-class ComparisonStrategy:
-    run_mode: RunMode = "parallel"
-
-    def __init__(
-        self,
-        fallback_provider: str | None = None,
-        on_mismatch: ComparisonMismatchHandler | None = None,
-    ) -> None:
-        self.fallback_provider = fallback_provider
-        self.on_mismatch = on_mismatch
-
-    def validate_provider_names(self, provider_names: Sequence[str]) -> None:
-        if (
-            self.fallback_provider is not None
-            and self.fallback_provider not in provider_names
-        ):
-            raise ValueError(
-                f"Fallback provider '{self.fallback_provider}' is not registered"
-            )
-
-    def should_use_result(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool:
-        del flag_key
-        del provider_name
-        del result
-        return False
-
-    def should_continue(
-        self,
-        flag_key: str,
-        provider_name: str,
-        result: FlagResolutionDetails[FlagValueType],
-    ) -> bool:
-        del flag_key
-        del provider_name
-        del result
-        return True
-
-    def determine_final_result(
-        self,
-        flag_key: str,
-        default_value: FlagValueType,
-        evaluations: list[_ProviderEvaluation[FlagValueType]],
-    ) -> FlagResolutionDetails[FlagValueType]:
-        """Determine the final result from parallel provider evaluations.
-
-        If any provider returns an error, returns an aggregated error result.
-        On agreement (all providers return the same value), returns the first
-        provider's result. On mismatch, calls the optional ``on_mismatch``
-        callback and returns the fallback provider's result.
-        """
-        if not evaluations:
-            return _build_aggregated_error(
-                flag_key,
-                default_value,
-                [],
-                "No providers were eligible for evaluation",
-            )
-        failed_evaluations = [
-            evaluation
-            for evaluation in evaluations
-            if not _is_success(evaluation.result)
-        ]
-        if failed_evaluations:
-            return _build_aggregated_error(
-                flag_key,
-                default_value,
-                failed_evaluations,
-                "Comparison strategy received provider errors",
-            )
-
-        # The first provider's result is the "final resolution" (used on agreement).
-        # The fallback provider's result is used on mismatch (per JS SDK reference).
-        final_evaluation = evaluations[0]
-        reference_value = final_evaluation.result.value
-        if not any(
-            evaluation.result.value != reference_value for evaluation in evaluations
-        ):
-            return final_evaluation.result
-
-        fallback_evaluation = self._select_fallback_evaluation(evaluations)
-        if self.on_mismatch is not None:
-            mismatch_results = {
-                evaluation.provider_name: evaluation.result
-                for evaluation in evaluations
-            }
-            try:
-                self.on_mismatch(flag_key, mismatch_results)
-            except Exception:
-                logger.exception(
-                    "Comparison strategy mismatch callback failed for flag '%s'",
-                    flag_key,
-                )
-        return fallback_evaluation.result
-
-    def _select_fallback_evaluation(
-        self, evaluations: list[_ProviderEvaluation[FlagValueType]]
-    ) -> _ProviderEvaluation[FlagValueType]:
-        if not evaluations:
-            raise ValueError("ComparisonStrategy requires at least one provider")
-        if self.fallback_provider is None:
-            return evaluations[0]
-        for evaluation in evaluations:
-            if evaluation.provider_name == self.fallback_provider:
-                return evaluation
-        raise ValueError(
-            f"Fallback provider '{self.fallback_provider}' is not registered"
-        )
 
 
 class MultiProvider(AbstractProvider):
@@ -770,35 +481,18 @@ class MultiProvider(AbstractProvider):
             resolution = resolve_fn(provider, flag_key, default_value, resolved_context)
             flag_evaluation = resolution.to_flag_evaluation_details(flag_key)
             if err := flag_evaluation.get_exception():
-                error_hooks(
-                    flag_type,
-                    err,
-                    reversed_hook_contexts,
-                    runtime.hook_hints,
-                )
+                error_hooks(flag_type, err, reversed_hook_contexts, runtime.hook_hints)
                 return _ProviderEvaluation(
-                    provider_name=provider_name,
-                    provider=provider,
-                    result=resolution,
+                    provider_name=provider_name, provider=provider, result=resolution
                 )
             after_hooks(
-                flag_type,
-                flag_evaluation,
-                reversed_hook_contexts,
-                runtime.hook_hints,
+                flag_type, flag_evaluation, reversed_hook_contexts, runtime.hook_hints
             )
             return _ProviderEvaluation(
-                provider_name=provider_name,
-                provider=provider,
-                result=resolution,
+                provider_name=provider_name, provider=provider, result=resolution
             )
         except Exception as err:
-            error_hooks(
-                flag_type,
-                err,
-                reversed_hook_contexts,
-                runtime.hook_hints,
-            )
+            error_hooks(flag_type, err, reversed_hook_contexts, runtime.hook_hints)
             return _ProviderEvaluation(
                 provider_name=provider_name,
                 provider=provider,
@@ -806,10 +500,7 @@ class MultiProvider(AbstractProvider):
             )
         finally:
             after_all_hooks(
-                flag_type,
-                flag_evaluation,
-                reversed_hook_contexts,
-                runtime.hook_hints,
+                flag_type, flag_evaluation, reversed_hook_contexts, runtime.hook_hints
             )
 
     async def _evaluate_provider_async(  # noqa: PLR0913
@@ -861,35 +552,18 @@ class MultiProvider(AbstractProvider):
             )
             flag_evaluation = resolution.to_flag_evaluation_details(flag_key)
             if err := flag_evaluation.get_exception():
-                error_hooks(
-                    flag_type,
-                    err,
-                    reversed_hook_contexts,
-                    runtime.hook_hints,
-                )
+                error_hooks(flag_type, err, reversed_hook_contexts, runtime.hook_hints)
                 return _ProviderEvaluation(
-                    provider_name=provider_name,
-                    provider=provider,
-                    result=resolution,
+                    provider_name=provider_name, provider=provider, result=resolution
                 )
             after_hooks(
-                flag_type,
-                flag_evaluation,
-                reversed_hook_contexts,
-                runtime.hook_hints,
+                flag_type, flag_evaluation, reversed_hook_contexts, runtime.hook_hints
             )
             return _ProviderEvaluation(
-                provider_name=provider_name,
-                provider=provider,
-                result=resolution,
+                provider_name=provider_name, provider=provider, result=resolution
             )
         except Exception as err:
-            error_hooks(
-                flag_type,
-                err,
-                reversed_hook_contexts,
-                runtime.hook_hints,
-            )
+            error_hooks(flag_type, err, reversed_hook_contexts, runtime.hook_hints)
             return _ProviderEvaluation(
                 provider_name=provider_name,
                 provider=provider,
@@ -897,10 +571,7 @@ class MultiProvider(AbstractProvider):
             )
         finally:
             after_all_hooks(
-                flag_type,
-                flag_evaluation,
-                reversed_hook_contexts,
-                runtime.hook_hints,
+                flag_type, flag_evaluation, reversed_hook_contexts, runtime.hook_hints
             )
 
     def _evaluate_with_providers(
@@ -950,10 +621,7 @@ class MultiProvider(AbstractProvider):
                 self.strategy.determine_final_result(
                     flag_key,
                     default_value,
-                    typing.cast(
-                        list[_ProviderEvaluation[FlagValueType]],
-                        evaluations,
-                    ),
+                    typing.cast(list[_ProviderEvaluation[FlagValueType]], evaluations),
                 ),
             )
 
@@ -986,10 +654,7 @@ class MultiProvider(AbstractProvider):
             self.strategy.determine_final_result(
                 flag_key,
                 default_value,
-                typing.cast(
-                    list[_ProviderEvaluation[FlagValueType]],
-                    evaluations,
-                ),
+                typing.cast(list[_ProviderEvaluation[FlagValueType]], evaluations),
             ),
         )
 
@@ -1033,10 +698,7 @@ class MultiProvider(AbstractProvider):
                 self.strategy.determine_final_result(
                     flag_key,
                     default_value,
-                    typing.cast(
-                        list[_ProviderEvaluation[FlagValueType]],
-                        evaluations,
-                    ),
+                    typing.cast(list[_ProviderEvaluation[FlagValueType]], evaluations),
                 ),
             )
 
@@ -1069,10 +731,7 @@ class MultiProvider(AbstractProvider):
             self.strategy.determine_final_result(
                 flag_key,
                 default_value,
-                typing.cast(
-                    list[_ProviderEvaluation[FlagValueType]],
-                    evaluations,
-                ),
+                typing.cast(list[_ProviderEvaluation[FlagValueType]], evaluations),
             ),
         )
 
@@ -1107,13 +766,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_boolean_details(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_boolean_details(k, d, c),
         )
 
     async def resolve_boolean_details_async(
@@ -1127,13 +780,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_boolean_details_async(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_boolean_details_async(k, d, c),
         )
 
     def resolve_string_details(
@@ -1147,13 +794,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_string_details(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_string_details(k, d, c),
         )
 
     async def resolve_string_details_async(
@@ -1167,13 +808,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_string_details_async(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_string_details_async(k, d, c),
         )
 
     def resolve_integer_details(
@@ -1187,13 +822,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_integer_details(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_integer_details(k, d, c),
         )
 
     async def resolve_integer_details_async(
@@ -1207,13 +836,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_integer_details_async(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_integer_details_async(k, d, c),
         )
 
     def resolve_float_details(
@@ -1227,13 +850,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_float_details(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_float_details(k, d, c),
         )
 
     async def resolve_float_details_async(
@@ -1247,13 +864,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_float_details_async(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_float_details_async(k, d, c),
         )
 
     def resolve_object_details(
@@ -1267,13 +878,7 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_object_details(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_object_details(k, d, c),
         )
 
     async def resolve_object_details_async(
@@ -1287,11 +892,5 @@ class MultiProvider(AbstractProvider):
             flag_key,
             default_value,
             evaluation_context,
-            lambda provider, resolved_flag_key, resolved_default_value, resolved_context: (
-                provider.resolve_object_details_async(
-                    resolved_flag_key,
-                    resolved_default_value,
-                    resolved_context,
-                )
-            ),
+            lambda p, k, d, c: p.resolve_object_details_async(k, d, c),
         )
